REORG: Moved common classes into a shared library to enable reuse.
