+++ DataFileHeader/DataSetString.cs
@@ -17,7 +17,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace DataFileTool
+namespace DataFileHeader
 {
     public class DataSetString
     {
