+++ DataFileHeader/FileHeader.cs
@@ -18,7 +18,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace DataFileTool
+namespace DataFileHeader
 {
     public class FileHeader
     {
