+++ DataFileTool.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29806.167
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32630.192
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataFileTool", "DataFileTool\DataFileTool.csproj", "{E5F6990D-CB84-4D76-977B-852BA4DC8CC9}"
 EndProject
@@ -9,6 +9,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataFileHeader", "DataFileHeader\DataFileHeader.csproj", "{03305213-42F9-480B-B9D0-952736806682}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,6 +22,10 @@ Global
 		{E5F6990D-CB84-4D76-977B-852BA4DC8CC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E5F6990D-CB84-4D76-977B-852BA4DC8CC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E5F6990D-CB84-4D76-977B-852BA4DC8CC9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{03305213-42F9-480B-B9D0-952736806682}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{03305213-42F9-480B-B9D0-952736806682}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03305213-42F9-480B-B9D0-952736806682}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{03305213-42F9-480B-B9D0-952736806682}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
