+++ DataFileTool/DataFileTool.csproj
@@ -6,4 +6,8 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DataFileHeader\DataFileHeader.csproj" />
+  </ItemGroup>
+
 </Project>
